move test_download_metadata mock pypi index utilities to conftest.py
